requirements.txt
Django==2.2
django-tinymce==3.0.2
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1
